Only keep extra peers if they give us a new block
If some time has passed and we didn't see a tip update, we open a new
extra connection. We do this to prevent the case where our subnet is
somehow partitioned from the global network and we don't get new blocks.
The problem with the current code is that it doesn't destroy another
connection and always keep the new one. If let running for enough time,
florestad will create too many connections and take a lot of CPU.

This commit fixes this by adding a new rule for extra peers; we only
keep them if they have blocks we didn't have. If they don't have any
unknown blocks, we disconnect. If they do, we disconnect one peer, based
on how fast they announce a new block to us. Faster peers are preferable
instead of slow ones.